requirements.txt
google-generativeai
python-dotenv
Pillow
torch
transformers
diffusers
huggingface_hub
streamlit
datetime
accelerate